refactor: refactor theme provider and hooks
