+++ src/hooks.ts
@@ -1,5 +1,6 @@
-import { useState, useEffect } from 'react';
+import { useState, useEffect, useContext } from 'react';
 import DEFAULT_SETTINGS from '@/../src-tauri/src/app/settings.json';
+import { ThemeContext } from './components/theme-provider';
 import { readSettings, writeSettings } from './utils';
 
 const DEFAULT_SETTING_STRING = JSON.stringify(DEFAULT_SETTINGS, null, 2);
@@ -18,4 +19,11 @@ export function useSettings() {
   }, []);
 
   return { settings , setSettings };
+}
+
+export function useTheme() {
+  const context = useContext(ThemeContext);
+  if (context === undefined) throw new Error('useTheme must be used within a ThemeProvider');
+
+  return context;
 }
